MatchTest uses the new Cypher doc build system
- Completed NewMatchTest and renamed it MatchTest
- Added IDREF to Section in DSL
- Renamed NewPrettyGraphTest to PrettyGraphTest
